feat: 상세페이지 첫 이동 시 리뷰 페이징 처리

+++ src/main/java/com/spring/dto/PagingVO.java
@@ -22,7 +22,7 @@ public class PagingVO {
 	}
 	
 	public void calTotalPage(int totalCnt,int cntPerPage) {
-		setTotalPage((int)Math.ceil(totalCnt/cntPerPage));
+		setTotalPage((int)Math.ceil((double)totalCnt/(double)cntPerPage));
 	}
 	
 	public void calStartEndPage(int nowPage,int cntPage) {
